Move duplicate code in matching requests to MatchActionRequestModel
